split companies-router tests into seperate files
