+++ __tests__/companies-model_find_all_companies.spec.js
